LayoutTests/imported/w3c/web-platform-tests/svg/text/scripted/getstartpositionofchar-expected.txt

ABC
ABC
ABC
ABC

FAIL vertical-rl assert_equals: expected 110 but got 100
FAIL vertical-lr assert_equals: expected 110 but got 100
FAIL sideways-rl assert_equals: expected 110 but got 100
FAIL sideways-lr assert_equals: expected 90 but got 100
